+++ koin/generator/test/src/commonMain/kotlin/GeneratedDefinitionsTest.kt
@@ -12,15 +12,27 @@ import org.koin.core.module.Module
 import org.koin.core.qualifier.named
 import org.koin.core.scope.Scope
 
+/**
+ * @return Definition by key "sampleInfo"
+ */
 public val Scope.sampleInfo: Test<String>
   get() = get(named("sampleInfo"))
 
+/**
+ * @return Definition by key "sampleInfo"
+ */
 public val Koin.sampleInfo: Test<String>
   get() = get(named("sampleInfo"))
 
+/**
+ * Will register [definition] with [org.koin.core.module.Module.single] and key "sampleInfo"
+ */
 public fun Module.sampleInfoSingle(createdAtStart: Boolean = false,
     definition: Definition<Test<String>>): KoinDefinition<Test<String>> =
     single(named("sampleInfo"), createdAtStart = createdAtStart, definition = definition)
 
+/**
+ * Will register [definition] with [org.koin.core.module.Module.factory] and key "sampleInfo"
+ */
 public fun Module.sampleInfoFactory(definition: Definition<Test<String>>):
     KoinDefinition<Test<String>> = factory(named("sampleInfo"), definition = definition)