+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * `Repos`:
     * `Exposed`:
         * `CommonExposedRepo.selectByIds` uses `foldRight` by default instead of raw foreach
+* `Koin`:
+    * `Generator`:
+        * Module has been created
 
 ## 0.16.11